tests and docs
